Fix compilation of chessgame.lyx with only TL 2018
chess-workshop-symbols.sty is no longer (on an updated TL 2018)
included in the TL package "skak". This commit replaces the loading
of chess-workshop-symbols in the custom preamble with the loading of
the package chessfss.

chessfss is in TL 2016 and TL 2018, and the example file now
compiles with both TL installations.

Before this commit, compilation of chessgame.lyx was possible if
chess-workshop-symbols.sty were installed manually (i.e., outside of
TL), but it is preferred to keep dependencies to only TL if
possible.