ctests: restore thesis.lyx lyx2lyx tests
These tests failed because of (multiple instances of) the following
warning:

  Warning: a buffer should not have two parents!

This warning is expected, because, for example, chapter-1 includes
Bibliography.lyx and chapter-2 includes Bibliography.lyx. thesis.lyx
includes both chapter-1 and chapter-2, so both are loaded (but
hidden), and Bibliography.lyx is loaded (but hidden), and LyX then
sees that more than one buffer includes Bibliography.lyx

The consequence of ignoring this warning is that we will not catch
any future regression that triggers this warning when it should not.