Tab binding: outline-in before depth-increment
Same for BackTab. The outline-in was originally (31398779)
introduced to the command-sequence at the end. Probably it was
placed at the end to be conservative (i.e., so that it would only
change behavior where there was a no-op before).

This fixes #11576.

(cherry picked from commit 9ab9f2b1acb8bc1a40d5a69737b43d09b7f7a597)