Give error if undefined citations with Biblatex
This commit catches situations when one key in a multi-key citation
is not defined. This commit causes an error to be given, but the
name of the key that is undefined is not provided in the error
dialog.

This commit is consistent with bf99ece7.

For more information, and possible follow-up discussion (e.g., on
putting the key in the error dialog), see the following ML thread:

https://www.mail-archive.com/search?l=mid&q=20190908165644.qnz6xu5bm5eqiko6%40boogie